first part of the architecture ready
