+++ MCTS.py
@@ -6,6 +6,6 @@ class MCTS:
     self.sample_size = sample_size
     self.action_size = action_size
 
-  def generate_action(self, n_simulations):
+  def generate_action(self, n_simulations, observation):
     # PLACEHOLDER
-    return np.random.randn(self.action_size)
+    return np.random.rand(observation.shape[0], self.action_size)
