+++ config.py
@@ -21,7 +21,7 @@ PB_C_BASE = 19652
 PB_C_INIT = 1.25
 DISCOUNT = 0.97
 TRAINING_STEPS = 1e10
-OBSERVATION_SIZE = 7804
+OBSERVATION_SIZE = [184, 4, 7]
 OBSERVATION_TIME_STEPS = 1
 OBSERVATION_TIME_STEP_INTERVAL = 1
 INPUT_TENSOR_SHAPE = np.array([OBSERVATION_SIZE])
